feat: Makes hc_analysis and hc_metric modules inside of hc_analysis
