+++ libs/hc_analysis/Cargo.toml
@@ -6,8 +6,16 @@ license = "Apache-2.0"
 publish = false
 
 [dependencies]
+content_inspector = "0.2.4"
 hc_common = { path = "../hc_common" }
 hc_data = { path = "../hc_data" }
-hc_metric = { path = "../hc_metric" }
 hc_config = { path = "../hc_config" }
 hc_report = { path = "../hc_report" }
+hc_shell = { path = "../hc_shell" }
+maplit = "1.0.2"
+unicode-normalization = "0.1.19"
+unicode-segmentation = "1.9.0"
+walkdir = "2"
+
+[dev-dependencies]
+hc_common = { path = "../hc_common" }
