+++ libs/hc_analysis/src/metric/activity.rs
