+++ libs/hc_analysis/src/metric/affiliation.rs
