+++ libs/hc_analysis/src/metric/binary.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::binary_detector::detect_binary_files;
+use crate::metric::binary_detector::detect_binary_files;
 use crate::MetricProvider;
 use hc_common::{
 	error::Result,
