+++ libs/hc_analysis/src/metric/binary_detector/mod.rs
