+++ libs/hc_analysis/src/metric/commit_trust.rs
