+++ libs/hc_analysis/src/metric/contributor_trust.rs
