+++ libs/hc_analysis/src/metric/identity.rs
