+++ libs/hc_analysis/src/metric/math.rs
