+++ libs/hc_analysis/src/metric/module.rs
