+++ libs/hc_analysis/src/metric/module_contributors.rs
