+++ libs/hc_analysis/src/metric/typo.rs
