--- libs/hc_metric/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "hc_metric"
-version = "0.1.0"
-edition = "2021"
-license = "Apache-2.0"
-publish = false
-
-[dependencies]
-content_inspector = "0.2.4"
-hc_common = { path = "../hc_common" }
-hc_config = { path = "../hc_config" }
-hc_data = { path = "../hc_data" }
-hc_shell = { path = "../hc_shell" }
-maplit = "1.0.2"
-unicode-normalization = "0.1.19"
-unicode-segmentation = "1.9.0"
-walkdir = "2"
-
-[dev-dependencies]
-hc_common = { path = "../hc_common" }
